refactor: reduce duplications by setting dtype fingerprint property
The previous bug was present in MHFingerprintTransformer, because it redefined the _transform method. The only difference with the base class _transform method was the dtype of the fingerprint array.

This refactor will prevent similar bugs if other modifications are applied to the _transform method, to support additional features, like concatenations of arrays.