refactor: use better Version class to skip tests based on versions
NOTE: this include an additional mandatory dependency, which is very small and pure-python: the `packaging` package. It's a package from the Python Packaging Authority, and it's the best tool to handle version information. It wasn't strictly necessary for this simple version check. But in the future, more complex version checks may be needed, and this package will help a lot.