test: skip tests based on sklearn version
It's not Python < 3.8 that doesn't support the sklearn set_output API: it's a sklearn version < 1.2.
A sklearn version < 1.2 could be present in an environment with Python > 3.8.

NOTE: should we raise a warning if the get_feature_names methods are used with sklearn < 1.2? In theory that method could still be used manually, but it would not be used authomatically by the sklearn set_output machinery.