feat: do not prefix fingerprints with zeros
Having fingerprints column names with the same length is better for display purposes: fp_morgan_0001...fp_morgan_0200...fp_morgan_2048.

But this may make it harder to test different fingerprint configurations: the column names will be based on the number of bits!
For instance, if nBits is 512, the first fingerprint is called fp_morgan_001; if nBits is 1024, the first fingerprint is fp_morgan_0001.

Now we will always call the first fingerprint fp_morgan_1. This will simplify processing the transformer outputs in production workflows.

An helper method that returns the nice zero-prefixed column names is included.